+++ packages/react/src/components/controls/MediaDeviceSelect.tsx
@@ -1,67 +1,8 @@
 import * as React from 'react';
 import { useMaybeRoomContext } from '../../context';
-import { setupDeviceSelector, createMediaDeviceObserver, log } from '@livekit/components-core';
 import { mergeProps } from '../../utils';
-import type { LocalAudioTrack, LocalVideoTrack, Room } from 'livekit-client';
-import { useObservableState } from '../../hooks/internal/useObservableState';
-
-/** @public */
-export function useMediaDevices({ kind }: { kind: MediaDeviceKind }) {
-  const deviceObserver = React.useMemo(() => createMediaDeviceObserver(kind), [kind]);
-  const devices = useObservableState(deviceObserver, []);
-  return devices;
-}
-
-/** @public */
-export interface UseMediaDeviceSelectProps {
-  kind: MediaDeviceKind;
-  room?: Room;
-  track?: LocalAudioTrack | LocalVideoTrack;
-  /**
-   * this will call getUserMedia if the permissions are not yet given to enumerate the devices with device labels.
-   * in some browsers multiple calls to getUserMedia result in multiple permission prompts.
-   * It's generally advised only flip this to true, once a (preview) track has been acquired successfully with the
-   * appropriate permissions.
-   *
-   * @see {@link MediaDeviceMenu}
-   * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/enumerateDevices | MDN enumerateDevices}
-   */
-  requestPermissions?: boolean;
-}
-
-/** @public */
-export function useMediaDeviceSelect({
-  kind,
-  room,
-  track,
-  requestPermissions,
-}: UseMediaDeviceSelectProps) {
-  const roomContext = useMaybeRoomContext();
-  // List of all devices.
-  const deviceObserver = React.useMemo(
-    () => createMediaDeviceObserver(kind, requestPermissions),
-    [kind, requestPermissions],
-  );
-  const devices = useObservableState(deviceObserver, []);
-  // Active device management.
-  const [currentDeviceId, setCurrentDeviceId] = React.useState<string>('');
-  const { className, activeDeviceObservable, setActiveMediaDevice } = React.useMemo(
-    () => setupDeviceSelector(kind, room ?? roomContext, track),
-    [kind, room, roomContext, track],
-  );
-
-  React.useEffect(() => {
-    const listener = activeDeviceObservable.subscribe((deviceId) => {
-      log.info('setCurrentDeviceId', deviceId);
-      if (deviceId) setCurrentDeviceId(deviceId);
-    });
-    return () => {
-      listener?.unsubscribe();
-    };
-  }, [activeDeviceObservable]);
-
-  return { devices, className, activeDeviceId: currentDeviceId, setActiveMediaDevice };
-}
+import type { LocalAudioTrack, LocalVideoTrack } from 'livekit-client';
+import { useMediaDeviceSelect } from '../../hooks';
 
 /** @public */
 export interface MediaDeviceSelectProps extends React.HTMLAttributes<HTMLUListElement> {
