+++ packages/react/src/hooks/useMediaDevices.ts
@@ -1,0 +1,10 @@
+import React from 'react';
+import { useObservableState } from './internal';
+import { createMediaDeviceObserver } from '@livekit/components-core';
+
+/** @public */
+export function useMediaDevices({ kind }: { kind: MediaDeviceKind }) {
+  const deviceObserver = React.useMemo(() => createMediaDeviceObserver(kind), [kind]);
+  const devices = useObservableState(deviceObserver, []);
+  return devices;
+}
